Restructure mongo-join-builder package
