+++ .changeset/978beb61/changes.json
@@ -1,0 +1,10 @@
+{
+  "releases": [{ "name": "@keystone-alpha/mongo-join-builder", "type": "major" }],
+  "dependents": [
+    {
+      "name": "@keystone-alpha/adapter-mongoose",
+      "type": "patch",
+      "dependencies": ["@keystone-alpha/mongo-join-builder"]
+    }
+  ]
+}
