+++ .changeset/cd56535e/changes.json
@@ -1,0 +1,1 @@
+{ "releases": [{ "name": "@keystone-alpha/adapter-mongoose", "type": "patch" }], "dependents": [] }
