+++ .changeset/cd56535e/changes.md
@@ -1,0 +1,1 @@
+- Use the updated mongo-join-builder package API.
