+++ packages/mongo-join-builder/index.js
@@ -1,11 +1,3 @@
-const { queryParser } = require('./query-parser');
-const joinBuilder = require('./join-builder');
+const mongoJoinBuilder = require('./lib/index');
 
-module.exports = parserOptions => {
-  return async (query, aggregate) => {
-    const queryTree = queryParser(parserOptions, query);
-    const { pipeline, postQueryMutations } = joinBuilder(queryTree);
-    // Run the query against the given database and collection
-    return await aggregate(pipeline).then(postQueryMutations);
-  };
-};
+module.exports = { mongoJoinBuilder };
